Updated Graphical_Kanban, Handler, Kanban and Save classes
Graphical_Kanban:
  - Updated Generate_Cards to adapt he "for" to the cards dictionnary.
- Updated Add_Column to edit the Column_Label.set_markup to remove
whitespace between <b> and <big> which was creating whitespace in the
label, therefore making errors later when using the label as a key to
retrieve the column object.

Handler:
  - Updated On_Rename_Dialog_Save_Clicked:
- In "Rename_Column" case, edit the Column_Label.set_markup to remove
whitespace between <b> and <big> which was creating whitespace in the
label, therefore making errors later when using the label as a key to
retrieve the column object.
  - It now saves at the end of this signal.
- Updated the Rename_Column case to correctly edit all the elements
related to the column, the object and the graphical elements.
- Updated On_Edit_Card_Dialog_Save_Clicked to save on each
modification and implemented the "Add_Card" case.
- Updated On_Kanban_Combo_Box_Changed to change the name of the file
when it loads another kanban (to avoid overwriting the wrong one).
- Updated Connect_Column_Buttons to send a column_box to
On_Column_Edit_Clicked instead of a column_label.
- Updated On_Column_Edit_Clicked to use the column_box instead of the
column_label

Kanban:
  - Added Set_Column_Title.
- Updated the "for" in Set_Columns to use .values to move through the
values of the dictionnary.

Save:
  - Removed a useless comment and edited an incorrect one.